+++ lib/karafka/web/pro/ui/lib/policies/requests.rb
@@ -17,19 +17,16 @@ module Karafka
       module Ui
         module Lib
           module Policies
-            # Extra configuration for pro UI
-            class Config
-              extend ::Karafka::Core::Configurable
-
-              # Policies controller related to messages operations and visibility
-              setting :messages, default: Policies::Messages.new
-
-              # Policies controller related to all requests. It is a general one that is not
-              # granular but can be used to block completely certain pieces of the UI from
-              # accessing like explorer or any other as operates on the raw env level
-              setting :requests, default: Policies::Requests.new
-
-              configure
+            # Per request policies setup class. Allows for writing low level policies that can
+            # be used to block certain Web UI functionalities on the request level.
+            class Requests
+              # @param _env [Hash] rack env object that we can use to get request details
+              # @return [Boolean] should this request be allowed or not
+              # @note By default we do not limit anything in the Web UI, however particular
+              #   granular policies may limit things on their own.
+              def allow?(_env)
+                true
+              end
             end
           end
         end